requirements.txt
fastapi
uvicorn 
boto3
python-multipart
python-dotenv
# tensorflow
# matplotlib
# seaborn